requirements.txt
flask>=2.0
pymongo>=3.12
python-dotenv>=1.0
